Add BattleGrid
This is the first version of the battle grid. This grid doesn't do shit yet, but at least it's there.

Gotta make sure there's also a UI at the bottom, that it's wide enough, and that you can scroll through it.

No units exist yet; gotta make sure that they can be placed, moved, etc.