Add Mine and Builder
Add Builder. A single builder with enough money to build a mine is awarded at the start of the battle.
Mines are buildings that give money at the start of each turn.
The amount is initialized to 15, although it's part of the config.
Building is done with the overlay when a Unit is selected.